Add rich interaction renderer extraction
