+++ apps/core/src/app/bootstrap/channel-wiring.ts
@@ -81,9 +81,7 @@ import { createChannelMessageActionRouter } from './channel-message-action-route
 import { createChannelProgressSender } from './channel-progress-sender.js';
 import { hydrateChannelConversationContext } from './channel-wiring-conversation-context.js';
 export type { ChannelWiring } from './channel-wiring-types.js';
-
 const PROVIDER_INBOUND_LEASE_PREFIX = 'runtime:provider-inbound';
-
 export function createChannelWiring(
   app: RuntimeApp,
   deps: Partial<ChannelWiringDeps> = {},
@@ -101,7 +99,6 @@ export function createChannelWiring(
     runtimeSecrets: new EnvRuntimeSecretProvider(),
     ...deps,
   };
-
   const connectedChannels: ChannelAdapter[] = [];
   const connectedChannelLeases: RuntimeLease[] = [];
   let enqueueRetryTailRecovery: RetryTailRecoveryEnqueue | undefined;
@@ -120,7 +117,6 @@ export function createChannelWiring(
       return undefined;
     }
   };
-
   let currentRuntimeSettings: RuntimeSettings;
   function findBoundChannel(jid: string): ChannelAdapter | undefined {
     return findChannel(connectedChannels, jid);
@@ -211,7 +207,6 @@ export function createChannelWiring(
     options?: { providerInbound?: boolean },
   ): Promise<void> {
     currentRuntimeSettings = runtimeSettings;
-    // Inbound needs the live-turns flag AND a role that admits inbound.
     const inboundEnabled =
       runtimeSettings.runtime.liveTurns.enabled &&
       (options?.providerInbound ?? true);
@@ -536,7 +531,6 @@ export function createChannelWiring(
       });
       throw thrownError;
     }
-
     if (options.durability === 'required' && durableAttempt) {
       const ambiguousSentSettlementError =
         'Provider send succeeded but durable sent-status persistence failed. Delivery may already be visible and cannot be blindly retried.';
@@ -587,7 +581,6 @@ export function createChannelWiring(
         });
       }
     }
-
     try {
       await outboundOps?.storeMessage({
         ...baseMessage,
@@ -640,19 +633,16 @@ export function createChannelWiring(
     });
     return result;
   }
-
   function setRetryTailRecoveryEnqueue(
     enqueue: RetryTailRecoveryEnqueue | undefined,
   ): void {
     enqueueRetryTailRecovery = enqueue;
   }
-
   function setDurableOutboundAttemptFactory(
     factory: DurableOutboundAttemptFactory | undefined,
   ): void {
     durableOutboundAttemptFactory = factory;
   }
-
   async function sendStreamingChunk(
     jid: string,
     rawText: string,
@@ -670,7 +660,6 @@ export function createChannelWiring(
     if (provider?.canStreamToJid?.(jid) === false) return false;
     const text = stripInternalTagsPreserveWhitespace(rawText);
     if (!text && !options?.done) return false;
-
     const sink = asStreamingSink(channel);
     if (!sink) return false;
     return sink.sendStreamingChunk(jid, text, options);
@@ -682,7 +671,6 @@ export function createChannelWiring(
     const stateSink = asStreamingStateSink(channel);
     stateSink?.resetStreaming(jid);
   }
-
   async function setTyping(jid: string, isTyping: boolean): Promise<void> {
     const channel = findBoundChannel(jid);
     if (!channel) return;
@@ -690,7 +678,6 @@ export function createChannelWiring(
     if (!typingSink) return;
     await typingSink.setTyping(jid, isTyping);
   }
-
   async function addReaction(
     jid: string,
     messageRef: string,
@@ -702,14 +689,12 @@ export function createChannelWiring(
     if (!reactionSink) return;
     await reactionSink.addReaction(jid, messageRef, emoji);
   }
-
   async function syncGroups(force: boolean): Promise<void> {
     const syncSources = connectedChannels
       .map(asGroupDiscoverySource)
       .filter((source): source is GroupDiscoverySource => source !== undefined);
     await Promise.all(syncSources.map((source) => source.syncGroups(force)));
   }
-
   async function disconnectChannels(): Promise<void> {
     const drained = await persistenceQueue.waitForIdle(5_000);
     if (!drained) {
