+++ apps/core/src/app/bootstrap/runtime-app.ts
@@ -21,6 +21,7 @@ import { ConversationRoute, ThinkingOverride } from '../../domain/types.js';
 import { RemoteMcpDnsValidationCache } from '../../application/mcp/mcp-server-policy.js';
 import { createGroupProcessor } from '../../runtime/group-processing.js';
 import type { GroupProcessingDeps } from '../../runtime/group-processing-types.js';
+import { resolveAgentLockStatus } from '../../config/profiles.js';
 import { listAvailableGroups } from '../../runtime/group-registry.js';
 import { GroupQueue } from '../../runtime/group-queue.js';
 import { parseThreadQueueKey } from '../../shared/thread-queue-key.js';
@@ -567,6 +568,7 @@ export function createRuntimeApp(options: RuntimeAppOptions = {}): RuntimeApp {
       getRuntimeStorage().repositories.patternCandidates,
     getProactiveSurfacingRepository: () =>
       getRuntimeStorage().repositories.proactiveSurfacing,
+    getAgentLockStatus: resolveAgentLockStatus,
     getSkillRepository: () => getRuntimeStorage().repositories.skills,
     getMcpServerRepository: () => getRuntimeStorage().repositories.mcpServers,
     getCapabilitySecretRepository: () =>
