+++ apps/core/src/channels/rich-interaction.ts
@@ -3,7 +3,6 @@ import type {
   RichInteractionRequest,
 } from '../domain/types.js';
 import { RICH_INTERACTION_NATIVE_FALLBACK_TEXT } from '../domain/types.js';
-import { TEAMS_ADAPTIVE_CARD_CONTENT_TYPE } from './teams-cards.js';
 
 export const RICH_INTERACTION_FALLBACK_COPY =
   RICH_INTERACTION_NATIVE_FALLBACK_TEXT;
@@ -19,19 +18,19 @@ type RichDescriptor = InteractionDescriptor & {
   fallbackText?: string;
 };
 
-function descriptor(input: RichInteractionRequest): RichDescriptor {
+export function richDescriptor(input: RichInteractionRequest): RichDescriptor {
   return input.descriptor;
 }
 
 export function richFallbackText(input: RichInteractionRequest): string {
-  const item = descriptor(input);
+  const item = richDescriptor(input);
   return (
     item.rich?.fallbackText || item.fallbackText || item.body || item.title
   );
 }
 
 function textLines(input: RichInteractionRequest): string[] {
-  const item = descriptor(input);
+  const item = richDescriptor(input);
   const lines = [item.title, item.body].filter(Boolean) as string[];
   for (const detail of item.details ?? []) {
     lines.push(`${detail.label}: ${detail.value}`);
@@ -45,27 +44,27 @@ function textLines(input: RichInteractionRequest): string[] {
   return lines.length ? lines : [richFallbackText(input)];
 }
 
-function slackEscape(text: string): string {
+export function richSlackEscape(text: string): string {
   return text
     .replace(/&/g, '&amp;')
     .replace(/</g, '&lt;')
     .replace(/>/g, '&gt;');
 }
 
-function htmlEscape(text: string): string {
-  return slackEscape(text).replace(/"/g, '&quot;');
+export function richHtmlEscape(text: string): string {
+  return richSlackEscape(text).replace(/"/g, '&quot;');
 }
 
-function truncate(text: string, max: number): string {
+export function richTruncate(text: string, max: number): string {
   return text.length > max ? `${text.slice(0, max - 1)}…` : text;
 }
 
-function isForm(input: RichInteractionRequest): boolean {
-  return descriptor(input).rich?.kind === 'form';
+export function isRichForm(input: RichInteractionRequest): boolean {
+  return richDescriptor(input).rich?.kind === 'form';
 }
 
 function payloadLines(input: RichInteractionRequest): string[] {
-  const rich = descriptor(input).rich;
+  const rich = richDescriptor(input).rich;
   const payload = rich?.payload ?? {};
   switch (rich?.kind) {
     case 'status':
@@ -75,11 +74,11 @@ function payloadLines(input: RichInteractionRequest): string[] {
         typeof payload.body === 'string' ? payload.body : '',
       ].filter(Boolean);
     case 'facts':
-      return arrayItems(payload.facts)
+      return richArrayItems(payload.facts)
         .map((fact) => lineFromPair(fact, 'label', 'value'))
         .filter(Boolean);
     case 'list':
-      return arrayItems(payload.items)
+      return richArrayItems(payload.items)
         .map(
           (item) =>
             lineFromPair(item, 'text', 'detail') ||
@@ -91,12 +90,12 @@ function payloadLines(input: RichInteractionRequest): string[] {
     case 'form':
       return [
         RICH_INTERACTION_REQUIRED_FIELDS_COPY,
-        ...arrayItems(payload.fields)
+        ...richArrayItems(payload.fields)
           .map((field) => lineFromPair(field, 'label', 'type'))
           .filter(Boolean),
       ];
     case 'media':
-      return arrayItems(payload.items)
+      return richArrayItems(payload.items)
         .map((item) => {
           const label = item.caption || item.alt || item.mime_type || 'Media';
           return lineFromPair({ ...item, label }, 'label', 'url');
@@ -112,7 +111,7 @@ function payloadLines(input: RichInteractionRequest): string[] {
   }
 }
 
-function arrayItems(value: unknown): Record<string, unknown>[] {
+export function richArrayItems(value: unknown): Record<string, unknown>[] {
   return Array.isArray(value)
     ? value.filter(
         (item): item is Record<string, unknown> =>
@@ -137,8 +136,8 @@ function lineFromPair(
 }
 
 function tableLines(payload: Record<string, unknown>): string[] {
-  const columns = arrayItems(payload.columns);
-  const rows = arrayItems(payload.rows);
+  const columns = richArrayItems(payload.columns);
+  const rows = richArrayItems(payload.rows);
   const keys = columns
     .map((column) => column.key)
     .filter((key): key is string => typeof key === 'string');
@@ -153,298 +152,14 @@ function tableLines(payload: Record<string, unknown>): string[] {
   );
 }
 
-function richTextLines(input: RichInteractionRequest): string[] {
+export function richTextLines(input: RichInteractionRequest): string[] {
   return [...textLines(input), ...payloadLines(input)].filter(Boolean);
 }
 
-function formFields(input: RichInteractionRequest): Record<string, unknown>[] {
-  return isForm(input)
-    ? arrayItems(descriptor(input).rich?.payload.fields).slice(0, 5)
+export function richFormFields(
+  input: RichInteractionRequest,
+): Record<string, unknown>[] {
+  return isRichForm(input)
+    ? richArrayItems(richDescriptor(input).rich?.payload.fields).slice(0, 5)
     : [];
-}
-
-export function buildSlackRichInteractionBlocks(
-  input: RichInteractionRequest,
-): Array<Record<string, unknown>> {
-  const item = descriptor(input);
-  const richLines = richTextLines(input);
-  const blocks: Array<Record<string, unknown>> = [
-    {
-      type: 'header',
-      text: {
-        type: 'plain_text',
-        text: truncate(item.title, 150),
-        emoji: true,
-      },
-    },
-  ];
-  const body = richLines.slice(1).join('\n');
-  if (body) {
-    blocks.push({
-      type: 'section',
-      text: { type: 'mrkdwn', text: truncate(slackEscape(body), 2900) },
-    });
-  }
-  if (item.details?.length) {
-    blocks.push({
-      type: 'section',
-      fields: item.details.slice(0, 10).map((detail) => ({
-        type: 'mrkdwn',
-        text: `*${slackEscape(detail.label)}*\n${slackEscape(detail.value)}`,
-      })),
-    });
-  }
-  if (item.actions?.length) {
-    blocks.push({
-      type: 'actions',
-      elements: item.actions.slice(0, 5).map((action) => ({
-        type: 'button',
-        text: {
-          type: 'plain_text',
-          text: truncate(action.label, 75),
-          emoji: true,
-        },
-        action_id: `gantry_rich_${action.id}`,
-        value: JSON.stringify({ interactionId: item.id, actionId: action.id }),
-        style:
-          action.style === 'danger'
-            ? 'danger'
-            : action.style === 'primary'
-              ? 'primary'
-              : undefined,
-      })),
-    });
-  }
-  if (isForm(input)) {
-    blocks.push(
-      {
-        type: 'context',
-        elements: [
-          {
-            type: 'mrkdwn',
-            text: slackEscape(RICH_INTERACTION_REQUIRED_FIELDS_COPY),
-          },
-        ],
-      },
-      {
-        type: 'actions',
-        elements: [
-          {
-            type: 'button',
-            text: {
-              type: 'plain_text',
-              text: RICH_INTERACTION_OPEN_FORM_LABEL,
-              emoji: true,
-            },
-            action_id: 'gantry_rich_form_open',
-            value: item.id,
-            style: 'primary',
-          },
-        ],
-      },
-    );
-  }
-  return blocks;
-}
-
-export function buildSlackRichInteractionFormModal(
-  input: RichInteractionRequest,
-): Record<string, unknown> {
-  const item = descriptor(input);
-  return {
-    type: 'modal',
-    callback_id: `gantry_rich_form_${item.id}`,
-    title: {
-      type: 'plain_text',
-      text: truncate(item.title || RICH_INTERACTION_OPEN_FORM_LABEL, 24),
-    },
-    submit: { type: 'plain_text', text: RICH_INTERACTION_SUBMIT_LABEL },
-    close: { type: 'plain_text', text: RICH_INTERACTION_CANCEL_LABEL },
-    blocks: [
-      {
-        type: 'section',
-        text: {
-          type: 'mrkdwn',
-          text: slackEscape(RICH_INTERACTION_REQUIRED_FIELDS_COPY),
-        },
-      },
-    ],
-  };
-}
-
-export function renderTelegramRichInteractionHtml(
-  input: RichInteractionRequest,
-): { text: string; reply_markup?: Record<string, unknown> } {
-  const item = descriptor(input);
-  const text = richTextLines(input)
-    .map((line, index) =>
-      index === 0 ? `<b>${htmlEscape(line)}</b>` : htmlEscape(line),
-    )
-    .join('\n');
-  const actions = item.actions?.slice(0, 8).map((action) => ({
-    text: action.label,
-    callback_data: `rich:${item.id}:${action.id}`.slice(0, 64),
-  }));
-  return {
-    text,
-    ...(actions?.length
-      ? { reply_markup: { inline_keyboard: actions.map((action) => [action]) } }
-      : {}),
-  };
-}
-
-export function buildDiscordRichInteractionPayload(
-  input: RichInteractionRequest,
-): { content: string; embeds: unknown[]; components?: unknown[] } {
-  const item = descriptor(input);
-  const richLines = richTextLines(input);
-  const fields = item.details?.slice(0, 25).map((detail) => ({
-    name: truncate(detail.label, 256),
-    value: truncate(detail.value, 1024) || ' ',
-    inline: true,
-  }));
-  const components = item.actions?.length
-    ? [
-        {
-          type: 1,
-          components: item.actions.slice(0, 5).map((action) => ({
-            type: 2,
-            label: truncate(action.label, 80),
-            style:
-              action.style === 'danger'
-                ? 4
-                : action.style === 'primary'
-                  ? 1
-                  : 2,
-            custom_id: `gantry:rich:${item.id}:${action.id}`.slice(0, 100),
-          })),
-        },
-      ]
-    : undefined;
-  const formComponents = isForm(input)
-    ? [
-        {
-          type: 1,
-          components: [
-            {
-              type: 2,
-              label: RICH_INTERACTION_OPEN_FORM_LABEL,
-              style: 1,
-              custom_id: `gantry:rich_form_open:${item.id}`.slice(0, 100),
-            },
-          ],
-        },
-      ]
-    : undefined;
-  return {
-    content: '',
-    embeds: [
-      {
-        title: truncate(item.title, 256),
-        description: truncate(
-          richLines.slice(1).join('\n') || richFallbackText(input),
-          4096,
-        ),
-        fields,
-      },
-    ],
-    components: formComponents ?? components,
-  };
-}
-
-export function buildDiscordRichInteractionFormModalResponse(
-  input: RichInteractionRequest,
-): Record<string, unknown> {
-  const item = descriptor(input);
-  const fields = arrayItems(item.rich?.payload.fields).slice(0, 5);
-  return {
-    type: 9,
-    data: {
-      custom_id: `gantry:rich_form_submit:${item.id}`.slice(0, 100),
-      title: truncate(item.title || RICH_INTERACTION_OPEN_FORM_LABEL, 45),
-      components: (fields.length
-        ? fields
-        : [{ label: RICH_INTERACTION_OPEN_FORM_LABEL }]
-      ).map((field, index) => ({
-        type: 1,
-        components: [
-          {
-            type: 4,
-            custom_id: `field_${index}`,
-            label: truncate(
-              String(field.label || field.id || `Field ${index + 1}`),
-              45,
-            ),
-            style: field.type === 'textarea' ? 2 : 1,
-            required: field.required === true,
-          },
-        ],
-      })),
-    },
-  };
-}
-
-export function buildTeamsRichInteractionPayload(
-  input: RichInteractionRequest,
-): {
-  attachments: [{ contentType: string; content: Record<string, unknown> }];
-} {
-  const item = descriptor(input);
-  const fields = formFields(input);
-  const body: Record<string, unknown>[] = richTextLines(input).map(
-    (line, index) => ({
-      type: 'TextBlock',
-      text: line,
-      wrap: true,
-      ...(index === 0 ? { size: 'Medium', weight: 'Bolder' } : {}),
-    }),
-  );
-  if (fields.length) {
-    body.push(
-      ...fields.map((field, index) => ({
-        type: 'Input.Text',
-        id: String(field.id || `field_${index}`),
-        label: String(field.label || field.id || `Field ${index + 1}`),
-        isMultiline: field.type === 'textarea',
-        isRequired: field.required === true,
-      })),
-    );
-  }
-  const actions: Record<string, unknown>[] = (item.actions ?? [])
-    .slice(0, 5)
-    .map((action) => ({
-      type: action.kind === 'submit' ? 'Action.Submit' : 'Action.Execute',
-      title: action.label,
-      verb: 'gantry.rich.action',
-      data: {
-        action: 'rich_interaction',
-        interactionId: item.id,
-        actionId: action.id,
-      },
-    }));
-  if (fields.length) {
-    actions.push({
-      type: 'Action.Submit',
-      title: RICH_INTERACTION_SUBMIT_LABEL,
-      verb: 'gantry.rich.form.submit',
-      data: {
-        action: 'rich_form_submit',
-        interactionId: item.id,
-      },
-    });
-  }
-  return {
-    attachments: [
-      {
-        contentType: TEAMS_ADAPTIVE_CARD_CONTENT_TYPE,
-        content: {
-          $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
-          type: 'AdaptiveCard',
-          version: '1.5',
-          body,
-          actions,
-        },
-      },
-    ],
-  };
 }
