+++ docs/decisions/README.md
@@ -5,34 +5,34 @@ the docs disagree.
 
 ## Current Decisions
 
-- [Runtime Stack Decision](0000-runtime-stack.md)
-- [Credential Boundary](0000-credential-broker-boundary.md)
-- [Agent Runtime Platform](0001-agent-runtime-platform.md)
-- [Config and Secret Source Boundary](0001-config-secret-source-boundary.md)
-- [Settings And Runtime Truth](2026-04-17-settings-runtime-truth.md)
-- [App-Wide Storage Backend Cutover](2026-04-21-storage-backend-cutover.md)
-- [Canonical Domain Schema Cutover](2026-04-26-canonical-domain-schema-cutover.md)
-- [Provider Artifact Store](2026-04-27-provider-session-artifact-store.md)
-- [Claude Runtime Materialization](2026-04-27-claude-runtime-materialization.md)
-- [Runtime Event Exchange](2026-04-29-runtime-event-exchange.md)
-- [Browser Capability Boundary](2026-04-29-browser-capability-boundary.md)
-- [External Ingress vs Outbound Webhooks](2026-04-30-external-ingress-vs-outbound-webhooks.md)
-- [Model Catalog and Cache Accounting](2026-05-01-model-catalog-and-cache-accounting.md)
-- [Event Bus Outbox Boundary](2026-05-12-event-bus-outbox-boundary.md)
-- [Provider-Neutral Agent Execution Adapter](2026-05-18-provider-neutral-agent-execution-adapter.md)
-- [JSONB Runtime Payload Boundary](2026-05-18-jsonb-runtime-payload-boundary.md)
-- [Simple Permission And Job Tool Lifecycle](2026-05-20-simple-permission-and-job-tool-lifecycle.md)
-- [MCP Source Vs Action Capability](2026-06-02-mcp-source-vs-action-capability.md)
-- [Settings Authority Per Deployment Mode](2026-06-11-settings-authority.md)
-- [Capability State As Artifacts](2026-06-11-capability-artifacts.md)
-- [Locked Agent Preset](2026-06-11-locked-preset.md)
-- [Delivery Vehicle](2026-06-11-delivery-vehicle.md)
-- [Deployment Modes](2026-06-11-deployment-modes.md)
-- [Agent Engine Selection](2026-06-12-agent-engine-selection.md)
-- [Process Roles and Multi-Live Execution](2026-06-12-process-roles-and-multi-live.md)
-- [Agent Harness Selection](2026-06-14-agent-harness-selection.md)
-- [Agent Communication Reaction Binding](2026-06-23-agent-communication-reaction-binding.md)
-- [Agent Communication Reasoning Safety](2026-06-23-agent-communication-reasoning-safety.md)
-- [Teams Reactions Deferred](2026-06-23-teams-reactions-deferred.md)
-- [Send Message Files Authority](2026-06-23-send-message-files-authority.md)
-- [Signed Artifact Links Deferred](2026-06-23-signed-artifact-links-deferred.md)
+- [Runtime Stack Decision](./0000-runtime-stack.md)
+- [Credential Boundary](./0000-credential-broker-boundary.md)
+- [Agent Runtime Platform](./0001-agent-runtime-platform.md)
+- [Config and Secret Source Boundary](./0001-config-secret-source-boundary.md)
+- [Settings And Runtime Truth](./2026-04-17-settings-runtime-truth.md)
+- [App-Wide Storage Backend Cutover](./2026-04-21-storage-backend-cutover.md)
+- [Canonical Domain Schema Cutover](./2026-04-26-canonical-domain-schema-cutover.md)
+- [Provider Artifact Store](./2026-04-27-provider-session-artifact-store.md)
+- [Claude Runtime Materialization](./2026-04-27-claude-runtime-materialization.md)
+- [Runtime Event Exchange](./2026-04-29-runtime-event-exchange.md)
+- [Browser Capability Boundary](./2026-04-29-browser-capability-boundary.md)
+- [External Ingress vs Outbound Webhooks](./2026-04-30-external-ingress-vs-outbound-webhooks.md)
+- [Model Catalog and Cache Accounting](./2026-05-01-model-catalog-and-cache-accounting.md)
+- [Event Bus Outbox Boundary](./2026-05-12-event-bus-outbox-boundary.md)
+- [Provider-Neutral Agent Execution Adapter](./2026-05-18-provider-neutral-agent-execution-adapter.md)
+- [JSONB Runtime Payload Boundary](./2026-05-18-jsonb-runtime-payload-boundary.md)
+- [Simple Permission And Job Tool Lifecycle](./2026-05-20-simple-permission-and-job-tool-lifecycle.md)
+- [MCP Source Vs Action Capability](./2026-06-02-mcp-source-vs-action-capability.md)
+- [Settings Authority Per Deployment Mode](./2026-06-11-settings-authority.md)
+- [Capability State As Artifacts](./2026-06-11-capability-artifacts.md)
+- [Locked Agent Preset](./2026-06-11-locked-preset.md)
+- [Delivery Vehicle](./2026-06-11-delivery-vehicle.md)
+- [Deployment Modes](./2026-06-11-deployment-modes.md)
+- [Agent Engine Selection](./2026-06-12-agent-engine-selection.md)
+- [Process Roles and Multi-Live Execution](./2026-06-12-process-roles-and-multi-live.md)
+- [Agent Harness Selection](./2026-06-14-agent-harness-selection.md)
+- [Agent Communication Reaction Binding](./2026-06-23-agent-communication-reaction-binding.md)
+- [Agent Communication Reasoning Safety](./2026-06-23-agent-communication-reasoning-safety.md)
+- [Teams Reactions Deferred](./2026-06-23-teams-reactions-deferred.md)
+- [Send Message Files Authority](./2026-06-23-send-message-files-authority.md)
+- [Signed Artifact Links Deferred](./2026-06-23-signed-artifact-links-deferred.md)
