+++ docs/product/README.md
@@ -3,7 +3,7 @@
 Product docs explain what Gantry is, who it serves, and which outcomes the
 runtime is built to protect.
 
-- [Product brief](BRIEF.md) - concise product intent, scope, and non-goals.
+- [Product brief](./BRIEF.md) - concise product intent, scope, and non-goals.
 - [Repository README](../../README.md) - installation, setup, and runtime overview.
 - [Requirements](../REQUIREMENTS.md) - product and architecture requirements.
 - [Specification](../SPEC.md) - expected runtime behavior and constraints.
